user authentication added

+++ src/components/ProjectDetails.js
@@ -1,6 +1,7 @@
 import { useState } from "react";
 import moment from "moment";
 import { useProjectsContext } from "../hooks/useProjectsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 import { currencyFormatter } from "../utilities/currencyFormatter";
 import ProjectForm from "./ProjectForm";
 
@@ -9,12 +10,18 @@ const ProjectDetails = ({ project }) => {
   const [isOverlayOpen, setIsOverlayOpen] = useState(false);
 
   const { dispatch } = useProjectsContext();
+  const { user } = useAuthContext();
 
   const handleDelete = async () => {
+    if (!user) {
+      return;
+    }
+
     const res = await fetch(
       `http://localhost:5000/api/projects/${project._id}`,
       {
         method: "DELETE",
+        Authorization: `Bearer ${user.token}`,
       }
     );
